src/third_party/WebKit/LayoutTests/http/tests/fetch/chromium/call-extra-crash-construct-and-look.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/stack-overflow.js"></script>
<script>
test(() => {
  let response;
  fillStackAndRun(() => response = new Response('hi'));
  assert_equals(response.body, undefined, 'response body should be undefined');
}, 'stack overflow in Response constructor should lead to an undefined body');
</script>
